Refactor fetch hooks
